urdf_property: refactor how TreeWidget is populated.
This changes the way the 'Robot Model' TreeView is populated and how
links and joints are displayed.

To not make navigating the TreeWidget too involved, the separation
between joints and links is kept, but instead of a flat list, the
hierarchical structure of both links and joints is maintained and
reflected in the organisation of the associated TreeItems.

TODO: refactor further and migrate to an MVC approach with a TreeView
and a Qt Data Model.