Refactor: reduce (include) dependencies.
This is a significant refactoring, which is intended to reduce the nr of include
dependencies in all source and header files. Forward declaration of classes is
used as much as possible, and includes have been moved to the cpp files as much
as possible as well.

This greatly increases locality of changes, resulting in shorter compilaton
times whenever changes have been made to any of the class definitions as it
avoids rippling significantly. Initial compilation is slightly slower, but that
is considered an acceptable trade-off.