travis: use default notification settings (for now).
According to [1], this will notify "committer and commit author" (all
committers are members of 'cad-to-ros pull', so have access to the
repository).

Did not specify a specific email address, as that would mean committer and
commit author don't get mailed anymore.

Quote from [1]:

  "By default, email notifications are sent to the committer and the
commit author, if they are members of the repository (that is, they have
push or admin permissions for public repositories, or if they have pull,
push or admin permissions for private repositories)."

[1] https://docs.travis-ci.com/user/notifications/#Notifications