Allow hiding/showing of model links.
These changes allow the MyRViz class to observe link visiblity changes
broadcast by the LinkProperty instances through a newly added
visibilityChanged(..) signal. The signal is emitted whenever the state of the
checkbox in the property browser widget is changed.

Link visibility is implemented as a 'pseudo-property', ie: it is not a
property of the actual model link, but only a visualisation setting. It
is treated like a property however to allow re-using the property browser
and the rest of the application infrastructure.

TODO: the slot implementation in MyRViz assumes that the link passed as
part of the signal exists and the RobotState display contains a reference
to it. No checking is implemented currently to make sure this is actually
true.

This is an adaptation of earlier work by Adam Allevato (@Kukanani).